xTaskCreateStatic call fixed

+++ d/freertos_druntime_backend/external/libc/config.d
@@ -1,4 +1,4 @@
 module external.libc.config;
 
 alias c_long = int;
-alias c_ulong = ulong;
+alias c_ulong = uint;
